Module description
In this no official Module you have to escape from a ghost that wants to kill you it doesn't matters what happens! it's gonna stalk you to kill and throw cannons towards you.

And watch out, because it may kill you with its body, so keep yourself at bay it!

The script is ran alone, it's only you start it with /lua.

Requirements: to play the game it's necessary exist at least 2 players playing in the room.

The script is gonna be updated soon. Only it, have fun!